ReSelect Proper Drive, Remove Optimized file from fileslist.
The proper drive needed to be re-selected before taking any defrag/optimize actions on the list of fragmented files, just in case the user had clicked off it, the defrag operation would fail as the volume to process was being based off Tab-1, not Tab-2
Made code to remove the defragmented/optimized file getting removed from the list after the defrag job. This way its obvious when the job completes and you dont have stale entries.